Add sidebar asset node menu to asset graph node as context menu (#17992)
## Summary & Motivation
Refactors the existing Sidebar popover menu so that it can be reused by
the asset graph itself for a context menu that appears on the individual
asset graph nodes when you right click.


## How I Tested These Changes
locally right click on a node in the asset graph and see the menu appear

<img width="603" alt="Screenshot 2023-11-14 at 2 46 43 PM"
src="https://github.com/dagster-io/dagster/assets/2286579/7f2087ca-1a96-487d-83cf-5feb3f226a4d">
