+++ js_modules/dagster-ui/packages/ui-core/src/asset-graph/AssetGraphExplorer.tsx
@@ -41,7 +41,7 @@ import {Loading} from '../ui/Loading';
 import {AssetEdges} from './AssetEdges';
 import {AssetGraphJobSidebar} from './AssetGraphJobSidebar';
 import {AssetGroupNode} from './AssetGroupNode';
-import {AssetNode, AssetNodeMinimal} from './AssetNode';
+import {AssetNode, AssetNodeMinimal, AssetNodeContextMenuWrapper} from './AssetNode';
 import {CollapsedGroupNode} from './CollapsedGroupNode';
 import {ExpandedGroupNode} from './ExpandedGroupNode';
 import {AssetNodeLink} from './ForeignNode';
@@ -487,6 +487,13 @@ const AssetGraphExplorerWithData = ({
                         return;
                       }
 
+                      const contextMenuProps = {
+                        graphData: fullAssetGraphData,
+                        node: graphNode,
+                        explorerPath,
+                        onChangeExplorerPath,
+                        selectNode: selectNodeById,
+                      };
                       return (
                         <foreignObject
                           {...bounds}
@@ -503,15 +510,19 @@ const AssetGraphExplorerWithData = ({
                           {!graphNode ? (
                             <AssetNodeLink assetKey={{path}} />
                           ) : scale < MINIMAL_SCALE ? (
-                            <AssetNodeMinimal
-                              definition={graphNode.definition}
-                              selected={selectedGraphNodes.includes(graphNode)}
-                            />
+                            <AssetNodeContextMenuWrapper {...contextMenuProps}>
+                              <AssetNodeMinimal
+                                definition={graphNode.definition}
+                                selected={selectedGraphNodes.includes(graphNode)}
+                              />
+                            </AssetNodeContextMenuWrapper>
                           ) : (
-                            <AssetNode
-                              definition={graphNode.definition}
-                              selected={selectedGraphNodes.includes(graphNode)}
-                            />
+                            <AssetNodeContextMenuWrapper {...contextMenuProps}>
+                              <AssetNode
+                                definition={graphNode.definition}
+                                selected={selectedGraphNodes.includes(graphNode)}
+                              />
+                            </AssetNodeContextMenuWrapper>
                           )}
                         </foreignObject>
                       );
