+++ js_modules/dagster-ui/packages/ui-core/src/asset-graph/AssetNode.tsx
@@ -3,6 +3,7 @@ import {Box, Colors, FontFamily, Icon, Spinner, Tooltip} from '@dagster-io/ui-co
 import countBy from 'lodash/countBy';
 import isEqual from 'lodash/isEqual';
 import React from 'react';
+import ReactDOM from 'react-dom';
 import {Link} from 'react-router-dom';
 import styled, {CSSObject} from 'styled-components';
 
@@ -13,11 +14,13 @@ import {StaleReasonsTags} from '../assets/Stale';
 import {assetDetailsPathForKey} from '../assets/assetDetailsPathForKey';
 import {AssetComputeKindTag} from '../graph/OpTags';
 import {AssetCheckExecutionResolvedStatus, AssetCheckSeverity} from '../graphql/types';
+import {ExplorerPath} from '../pipelines/PipelinePathUtils';
 import {markdownToPlaintext} from '../ui/markdownToPlaintext';
 
+import {useAssetNodeMenu} from './AssetNodeMenu';
 import {buildAssetNodeStatusContent} from './AssetNodeStatusContent';
 import {AssetLatestRunSpinner} from './AssetRunLinking';
-import {LiveDataForNode} from './Utils';
+import {GraphData, GraphNode, LiveDataForNode} from './Utils';
 import {ASSET_NODE_NAME_MAX_LENGTH} from './layout';
 import {AssetNodeFragment} from './types/AssetNode.types';
 
@@ -160,6 +163,100 @@ const AssetCheckIconsOrdered: {type: AssetCheckIconType; content: React.ReactNod
     content: <Icon name="check_circle" color={Colors.Green700} />,
   },
 ];
+
+export const AssetNodeContextMenuWrapper = React.memo(
+  ({
+    children,
+    graphData,
+    explorerPath,
+    onChangeExplorerPath,
+    selectNode,
+    node,
+  }: {
+    children: React.ReactNode;
+    graphData: GraphData;
+    node: GraphNode;
+    selectNode?: (e: React.MouseEvent<any> | React.KeyboardEvent<any>, nodeId: string) => void;
+    explorerPath?: ExplorerPath;
+    onChangeExplorerPath?: (path: ExplorerPath, mode: 'replace' | 'push') => void;
+  }) => {
+    const [menuVisible, setMenuVisible] = React.useState(false);
+    const [menuPosition, setMenuPosition] = React.useState<{top: number; left: number}>({
+      top: 0,
+      left: 0,
+    });
+
+    const showMenu = (e: React.MouseEvent) => {
+      e.preventDefault();
+      setMenuVisible(true);
+      setMenuPosition({top: e.pageY, left: e.pageX});
+    };
+
+    const hideMenu = () => {
+      setMenuVisible(false);
+    };
+    const ref = React.useRef<HTMLDivElement | null>(null);
+    React.useEffect(() => {
+      const node = ref.current;
+      const listener = (e: MouseEvent) => {
+        if (ref.current && e.target && !ref.current.contains(e.target as Node)) {
+          hideMenu();
+        }
+      };
+      const keydownListener = (e: KeyboardEvent) => {
+        if (ref.current && e.code === 'Escape') {
+          hideMenu();
+        }
+      };
+      if (menuVisible && node) {
+        document.body.addEventListener('click', listener);
+        document.body.addEventListener('keydown', keydownListener);
+      }
+      return () => {
+        if (node) {
+          document.body.removeEventListener('click', listener);
+          document.body.removeEventListener('keydown', keydownListener);
+        }
+      };
+    }, [menuVisible]);
+    const {dialog, menu} = useAssetNodeMenu({
+      graphData,
+      explorerPath,
+      onChangeExplorerPath,
+      selectNode,
+      node,
+    });
+    return (
+      <div ref={ref}>
+        <div onContextMenu={showMenu} onClick={hideMenu}>
+          {children}
+        </div>
+        {dialog}
+        {menuVisible
+          ? ReactDOM.createPortal(
+              <div
+                style={{
+                  position: 'absolute',
+                  top: menuPosition.top,
+                  left: menuPosition.left,
+                  backgroundColor: '#fff',
+                  border: '1px solid #ccc',
+                  boxShadow: '0 2px 4px rgba(0, 0, 0, 0.1)',
+                  zIndex: 10,
+                }}
+                onClick={(e) => {
+                  e.stopPropagation();
+                }}
+              >
+                {menu}
+              </div>,
+              document.body,
+            )
+          : null}
+      </div>
+    );
+  },
+);
 
 const AssetNodeChecksRow = ({
   definition,
