+++ js_modules/dagster-ui/packages/ui-core/src/assets/AssetNodeLineageGraph.tsx
@@ -8,7 +8,7 @@ import {useFeatureFlags} from '../app/Flags';
 import {AssetEdges} from '../asset-graph/AssetEdges';
 import {MINIMAL_SCALE} from '../asset-graph/AssetGraphExplorer';
 import {AssetGroupNode} from '../asset-graph/AssetGroupNode';
-import {AssetNodeMinimal, AssetNode} from '../asset-graph/AssetNode';
+import {AssetNodeMinimal, AssetNode, AssetNodeContextMenuWrapper} from '../asset-graph/AssetNode';
 import {ExpandedGroupNode} from '../asset-graph/ExpandedGroupNode';
 import {AssetNodeLink} from '../asset-graph/ForeignNode';
 import {GraphData, groupIdForNode, toGraphId} from '../asset-graph/Utils';
@@ -111,6 +111,11 @@ export const AssetNodeLineageGraph = ({
               const graphNode = assetGraphData.nodes[id];
               const path = JSON.parse(id);
 
+              const contextMenuProps = {
+                graphData: assetGraphData,
+                node: graphNode!,
+              };
+
               return (
                 <foreignObject
                   {...bounds}
@@ -127,15 +132,19 @@ export const AssetNodeLineageGraph = ({
                   {!graphNode ? (
                     <AssetNodeLink assetKey={{path}} />
                   ) : scale < MINIMAL_SCALE ? (
-                    <AssetNodeMinimal
-                      definition={graphNode.definition}
-                      selected={graphNode.id === assetGraphId}
-                    />
+                    <AssetNodeContextMenuWrapper {...contextMenuProps}>
+                      <AssetNodeMinimal
+                        definition={graphNode.definition}
+                        selected={graphNode.id === assetGraphId}
+                      />
+                    </AssetNodeContextMenuWrapper>
                   ) : (
-                    <AssetNode
-                      definition={graphNode.definition}
-                      selected={graphNode.id === assetGraphId}
-                    />
+                    <AssetNodeContextMenuWrapper {...contextMenuProps}>
+                      <AssetNode
+                        definition={graphNode.definition}
+                        selected={graphNode.id === assetGraphId}
+                      />
+                    </AssetNodeContextMenuWrapper>
                   )}
                 </foreignObject>
               );
